issue #41
The team creator is still being used as if it were the team leader role.
We will introduce the proper use of roles shortly.

The "creator" can be handed over to another team member. Also the team
creator can add new players. When the creator leaves the team his role
is assigned to 'another' team member.